First percona cluster implem

+++ cookbooks/mysql/recipes/server.rb
@@ -1,4 +1,7 @@
 
+require 'socket'
+require 'ipaddr'
+
 include_recipe "mysql"
 
 server_package_name = node.mysql.use_percona ? node.mysql.percona_server_package_name : node.mysql.server_package_name
@@ -41,10 +44,16 @@ file "/etc/mysql/conf.d/chef_override.cnf" do
   notifies :restart, "service[mysql]", :immediately
 end
 
-root_mysql_password = local_storage_read("mysql_password:root") do
-  password = PasswordGenerator.generate 32
-  Chef::Log.info "Mysql : new root password generated"
-  password
+root_mysql_password = node.mysql[:root_password]
+
+if root_mysql_password
+  local_storage_write("mysql_password:root", root_mysql_password)
+else
+  root_mysql_password = local_storage_read("mysql_password:root") do
+    password = PasswordGenerator.generate 32
+    Chef::Log.info "Mysql : new root password generated"
+    password
+  end
 end
 
 execute "change mysql root password" do
@@ -52,7 +61,47 @@ execute "change mysql root password" do
   only_if "echo 'select 1;' | mysql --user=root --password= "
 end
 
-if node[:mysql] && ! node.mysql[:keep_test]
+if node.mysql.use_percona && node.mysql[:percona_cluster]
+
+  local_node = nil
+  node["network"]["interfaces"].each do |name, config|
+    config["addresses"].each do |ip, ip_config|
+    local_node = ip if node.mysql.percona_cluster.nodes.include? ip
+    end
+  end
+
+  raise "No local ip found in node cluster definition #{node.mysql.percona_cluster.name}" unless local_node
+
+  config = node.mysql.percona_cluster.to_hash
+  config["local_node"] = local_node
+  config["root_password"] = root_mysql_password
+  config["is_master"] = local_node == config["master"]
+
+  ruby_block "update percona cluster" do
+    block do
+      PerconaCluster.new(config).converge true
+    end
+    action :nothing
+  end
+
+  template "/etc/mysql/conf.d/cluster.cnf" do
+    source 'percona_cluster.cnf.erb'
+    mode '0644'
+    variables config
+    notifies :run, "ruby_block[update percona cluster]", :immediately
+  end
+
+  ruby_block "manage percona cluster" do
+    block do
+      PerconaCluster.new(config).converge
+    end
+  end
+
+  node.set[:mysql][:run_sql] = false unless config["is_master"]
+
+end
+
+if node[:mysql] && ! node.mysql[:keep_test] && node.mysql[:run_sql]
   execute "remove test mysql database" do
     command "echo 'drop database test;' | mysql --user=root --password=#{root_mysql_password} "
     only_if "echo 'show databases;' | mysql --user=root --password=#{root_mysql_password} | grep ^test$ "
